requirements.txt
Flask>=0.10.1
leancloud-sdk>=1.0.9
pyopenssl
ndg-httpsclient
pyasn1
itsdangerous
